+++ relay-server/src/services/store.rs
@@ -772,6 +772,17 @@ impl StoreService {
                 );
                 return Ok(());
             }
+            MetricNamespace::Profiles => {
+                if !self
+                    .global_config
+                    .current()
+                    .options
+                    .profiles_function_generic_metrics_enabled
+                {
+                    return Ok(());
+                }
+                KafkaTopic::MetricsGeneric
+            }
             _ => KafkaTopic::MetricsGeneric,
         };
         let headers = BTreeMap::from([("namespace".to_string(), namespace.to_string())]);