+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Add SDK information to spans. ([#3178](https://github.com/getsentry/relay/pull/3178))
 - Filter null values from metrics summary tags. ([#3204](https://github.com/getsentry/relay/pull/3204))
 - Emit a usage metric for every span seen. ([#3209](https://github.com/getsentry/relay/pull/3209))
+- Add namespace for profile metrics. ([#3229](https://github.com/getsentry/relay/pull/3229))
 
 ## 24.2.0